moved some components to global level (heading, charts)
