+++ js/components/heading.html
@@ -1,0 +1,4 @@
+<div data-bind="css: classes()">
+  <i class="fa fa-area-chart"></i>
+  <span data-bind="text: name"></span> Report (<span data-bind="text: sourceKey"></span>)
+</div>
