+++ js/components/heading.less
@@ -1,0 +1,6 @@
+.heading-title {
+  font-size: 14px;
+  color: #000;
+  margin: 1.5rem 0 2rem 0;
+  font-weight: 500;
+}
