--- js/pages/data-sources/components/report-title.html
@@ -1,5 +1,0 @@
-<div class="reportHeading">
-  <i class="fa fa-area-chart"></i>
-  <span data-bind="text: $component.name"></span> Report (
-  <span data-bind="text: $component.sourceKey"></span>)
-</div>
