--- js/pages/data-sources/components/report-title.js
@@ -1,20 +1,0 @@
-define([
-	'knockout',
-	'text!./report-title.html',
-], function (
-	ko,
-	view,
-) {
-	function reportTitle(params) {
-    this.name = params.name;
-    this.sourceKey = params.sourceKey;
-  }
-
-	var component = {
-		viewModel: reportTitle,
-		template: view,
-	};
-
-	ko.components.register('report-title', component);
-	return component;
-});
