+++ js/pages/data-sources/components/reports/dashboard.js
@@ -5,12 +5,12 @@ define([
 	'atlascharts',
 	'd3-tip',
   'services/http',
-  'pages/data-sources/const',
+  'const',
   'pages/data-sources/classes/Report',
-  'pages/data-sources/components/report-title',
-  'pages/data-sources/components/charts/donut',
-  'pages/data-sources/components/charts/histogram',
-  'pages/data-sources/components/charts/line',
+  'components/heading',
+  'components/charts/donut',
+  'components/charts/histogram',
+  'components/charts/line',
 ], function (
 	ko,
 	view,
