+++ js/pages/data-sources/components/reports/datadensity.js
@@ -3,11 +3,11 @@ define([
 	'text!./datadensity.html',
 	'd3',
   'services/http',
-  'pages/data-sources/const',
+  'const',
   'pages/data-sources/classes/Report',
-  'pages/data-sources/components/report-title',
-  'pages/data-sources/components/charts/boxplot',
-  'pages/data-sources/components/charts/line',
+  'components/heading',
+  'components/charts/boxplot',
+  'components/charts/line',
 ], function (
 	ko,
 	view,
