+++ js/pages/data-sources/components/reports/death.html
@@ -1,5 +1,5 @@
 <!-- ko if: !isLoading() -->
-<report-title params="name: name, sourceKey: sourceKey"></report-title>
+<heading params="name: name, sourceKey: sourceKey"></heading>
 <div class="col-sm-12">
   <div class="panel panel-primary">
     <div class="panel-heading">
