+++ js/pages/data-sources/components/reports/visit.html
@@ -1,5 +1,5 @@
 <!-- ko if: !isLoading() -->
-<report-title params="name: name, sourceKey: sourceKey"></report-title>
+<heading params="name: name, sourceKey: sourceKey"></heading>
 <div class="panel panel-primary">
   <div class="panel-heading">
     <span data-bind="name"></span> Prevalence
@@ -15,7 +15,7 @@
     </ul>
     <div id='content' class="tab-content">
       <div role="tabpanel" class="tab-pane active" id="treemap-panel">
-        <treemap params="data: treeData, format: chartFormats.treemap"></treemap>
+        <treemap params="data: treeData, format: chartFormats.treemap, aggProperty: aggProperty"></treemap>
         <div class="treemap_legend">Box Size: Prevalence, Color: <span data-bind="text: aggProperty.description"></span> (Light to Dark = Low to High), Use Ctrl-Click to Zoom, Alt-Click to Reset Zoom
         </div>
       </div>
@@ -35,4 +35,7 @@
     </div>
   </div>
 </div>
+<!-- ko if: currentConcept() -->
+  <treemap-drilldown params="context: context, currentReport: currentReport, currentConcept: currentConcept"></treemap-drilldown>
+<!-- /ko -->
 <!--/ko-->
