+++ js/providers/Chart.js
@@ -1,8 +1,8 @@
 define([
   'knockout',
   'providers/Component',
-  'text!../components/charts/chart.html',
-  'pages/data-sources/const',
+  'text!components/charts/chart.html',
+  'const',
   'components/empty-state',
 ], function (
   ko,
@@ -15,8 +15,8 @@ define([
       super();
       this.view = view;
       this.container = ko.observable();
-      this.container.subscribe(this.draw.bind(this));
       this.data = ko.observable();
+      this.container.subscribe(this.draw.bind(this));
       this.data.subscribe(this.draw.bind(this));
       this.format = {};
       this.setContainer = this.setContainer.bind(this);
@@ -24,15 +24,15 @@ define([
 
     setContainer(element) {
       this.container(element);
+      this.width = this.container().getBoundingClientRect().width;
     }
     
     draw() {
       if (!this.container() || !this.data()) {
         return false;
       }
-      this.width = this.container().getBoundingClientRect().width;
       this.chart.render(
-        this.data(),
+        this.prepareData(this.data()),
         this.container(),
         this.width,
         this.minHeight,
@@ -40,10 +40,18 @@ define([
       );
     }
 
-    render(params) {
+    prepareData(rawData) {
+      return rawData;
+    }
+
+    storeParams(params) {
       this.minHeight = params.minHeight || constants.minChartHeight;
-      this.data(params.data());
       this.format = params.format;
+    }
+
+    render(params) {
+      this.storeParams(params);
+      this.data(params.data());
     }
   }
